trigger 294: objInterpretSeq/getLActions legCode is s8 + reachScale /16.0f (4 fns -> 100, unit 99.68->99.95)

+++ include/main/render_lactions_api.h
@@ -3,6 +3,6 @@
 
 #include "types.h"
 
-int getLActions(void* source, void* target, u16 index, int arg3, int arg4, int arg5);
+int getLActions(void* source, void* target, u16 index, s8 arg3, int arg4, int arg5);
 
 #endif /* MAIN_RENDER_LACTIONS_API_H_ */
